backend/requirements.txt
alembic==1.16.1
APScheduler==3.11.0
attrs==25.3.0
babel==2.17.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distlib==0.3.9
filelock==3.18.0
flasgger==0.9.7.1
Flask==3.1.1
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
google-auth==2.40.3
greenlet==3.2.3
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.1
mistune==3.1.3
packaging==25.0
pillow==11.3.0
pipenv==2025.0.3
platformdirs==4.3.8
pluggy==1.6.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.41
svgwrite==1.4.3
Tree==0.2.4
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
virtualenv==20.31.2
waitress==3.0.2
Werkzeug==3.1.3
